Add Chrome bookmarks in the Project.
Add permission for bookmarks in manifest.json, add new variable in the state and modify action and reducer for in changing.
Change header to split working and setting icons in differen angles of the screen. Add icon of Bookmarks in the Header.
Add Bookmarks component and container in the SlidingWindow. Add receiving of an array from the Chrome Bookmarks API. Show the basic list of bookmarks links and folders on the page.